requirements.txt
PyYAML>=6.0
requests>=2.28.0
slack-sdk>=3.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
